[Diagrams] Uploading class diagram files
Created class diagram files showing relationships between different
aspects of the program. The objects package contains the objects within
the program that reside in the database, such as Users, Items,
Ingredients, Orders, and the Inventory. Enums contain the type for
Ingredients used in Subs and the status of Orders placed by Customers.
Controllers contains all of the controllers that will be used to connect
all of the frontend user interactions to the backend/database objects.
Repositories and Services contain the list of all Objects and the
functions that can be used to retrieve them for manipulation. The main
package includes the application's main function.